% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_colours.R
\name{bj_cols}
\alias{bj_cols}
\title{Function to extract colors as hex codes}
\usage{
bj_cols(...)
}
\arguments{
\item{...}{Character names of bj_colors}
}
\description{
Function to extract colors as hex codes
}
