% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_plot.R
\name{gauge_plot}
\alias{gauge_plot}
\title{Gauge plots}
\usage{
gauge_plot(
  percentage,
  text_colour = "white",
  first_break = 0.5,
  second_break = 0.8,
  reverse = FALSE
)
}
\arguments{
\item{percentage}{Percentage to be plotted as a proportion between 0 and 1}

\item{text_colour}{Defaults to white for dark backgrounds}

\item{first_break}{Lower proportion colour break. Default is 0.5}

\item{second_break}{Upper proportion colour break. Default is 0.8}

\item{reverse}{Whether to reverse the colour scheme}
}
\description{
Create simple gauge plots using a proportion. Optionally select text colour, segment colour and order.
}
\examples{
\dontrun{
gauge_plot(0.35, text_colour = "black")
gauge_plot(0.85, text_colour = bj_cols("text grey"), reverse = TRUE)
gauge_plot(0.12, text_colour = "black", first_break = 0.1) }

}
